Change TileMapDrawer and SpriteDrawer constructor parameters from pointers to references
This better communicates that it is not valid to pass nullptr for these
parameters.

This also makes these classes not copyable or movable because copy and move are
not necessary.  If they become necessary then the members can be changed to
pointers but the constructor parameters should stay as references.