Make ShaderProgram accept vertex and fragment shaders as references.
const references are used because the shaders will not be modified and they are
only used for the duration of the constructor call.  Therefore it would be
valid to pass temporaries.

This removes the unnecessary initialization list of shaders that ShaderProgram
accepted before.  For right now and for the foreseeable future there will just
be a vertex shader and a fragment shader so taking them as separate explicit
arguments is simpler.