Clean up includes from Window library
Now the Window library publishes a proper include directory for other libraries
to use without depending on relative paths to the Window library header files.

We may want to organize other libraries this way in the future and possibly
even see if this strategy could help with splitting Graphics and GraphicsOpenGL
into separate libraries.