Write test test that fails if destructor is not called.
Also learned that this whole placement new strategy may be more complicated
than I want.  I am considering boost::object_pool or possibly just using a
vector with reserved space.  But I want to iterate the objects in sequential
order and also be able to de-allocate and re-use blocks in the middle.